# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAnonymi API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/documents/intake-defaults/{workspace_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Intake Defaults Route",
        "operationId": "intake_defaults_route_api_v1_documents_intake_defaults__workspace_id__get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeDefaultsView"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Create Pasted Draft Route",
        "operationId": "create_pasted_draft_route_api_v1_documents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedDraftView"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/from-file": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Create File Draft Route",
        "operationId": "create_file_draft_route_api_v1_documents_from_file_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_file_draft_route_api_v1_documents_from_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedDraftView"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/source": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Source Route",
        "operationId": "source_route_api_v1_documents__document_id__source_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceView"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "documents"
        ],
        "summary": "Edit Source Route",
        "operationId": "edit_source_route_api_v1_documents__document_id__source_put",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditSourceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedDraftView"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/revisions/{revision_id}/source": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Current Source Route",
        "operationId": "current_source_route_api_v1_documents__document_id__revisions__revision_id__source_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Revision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceView"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_file_draft_route_api_v1_documents_from_file_post": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "categories": {
            "type": "string",
            "title": "Categories",
            "default": "email,phone"
          },
          "phone_region": {
            "type": "string",
            "maxLength": 2,
            "minLength": 2,
            "title": "Phone Region",
            "default": "PH"
          },
          "retention_days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 30.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Retention Days"
          },
          "preset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset Id"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "file"
        ],
        "title": "Body_create_file_draft_route_api_v1_documents_from_file_post"
      },
...
      "ConflictResponse": {
        "properties": {
          "code": {
            "type": "string",
            "const": "version_conflict",
            "title": "Code",
            "default": "version_conflict"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "This review changed in another session. Reload before saving."
          },
          "current_version": {
            "$ref": "#/components/schemas/VersionRef"
          }
        },
        "type": "object",
        "required": [
          "current_version"
        ],
        "title": "ConflictResponse"
      },
...
      "CreateDraftRequest": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/FindingCategory"
            },
            "type": "array",
            "title": "Categories"
          },
          "phone_region": {
            "type": "string",
            "maxLength": 2,
            "minLength": 2,
            "title": "Phone Region",
            "default": "PH"
          },
          "retention_days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 30.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Retention Days"
          },
          "preset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset Id"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "source"
        ],
        "title": "CreateDraftRequest"
      },
...
      "DocumentStatus": {
        "type": "string",
        "enum": [
          "draft",
          "scanning",
          "needs_review",
          "ready",
          "exported",
          "failed",
          "expired",
          "deleted"
        ],
        "title": "DocumentStatus"
      },
      "EditSourceRequest": {
        "properties": {
          "expected": {
            "$ref": "#/components/schemas/VersionRef"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "expected",
          "source"
        ],
        "title": "EditSourceRequest"
      },
      "ErrorDetail": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "issue": {
            "type": "string",
            "title": "Issue"
          }
        },
        "type": "object",
        "required": [
          "field",
          "issue"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "FindingCategory": {
        "type": "string",
        "enum": [
          "person",
          "organization",
          "location",
          "address",
          "identifier",
          "custom",
          "email",
          "phone"
        ],
        "title": "FindingCategory"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IntakeDefaultsView": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "content_retention_days": {
            "type": "integer",
            "title": "Content Retention Days"
          },
          "current_time": {
            "type": "string",
            "format": "date-time",
            "title": "Current Time"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "content_retention_days",
          "current_time"
        ],
        "title": "IntakeDefaultsView"
      },
...
      "SavedDraftView": {
        "properties": {
          "version": {
            "$ref": "#/components/schemas/VersionRef"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          }
        },
        "type": "object",
        "required": [
          "version",
          "expires_at",
          "status"
        ],
        "title": "SavedDraftView"
      },
...
      "SourceView": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workspace Id"
          },
          "can_edit": {
            "type": "boolean",
            "title": "Can Edit"
          },
          "version": {
            "$ref": "#/components/schemas/VersionRef"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/FindingCategory"
            },
            "type": "array",
            "title": "Categories"
          },
          "phone_region": {
            "type": "string",
            "title": "Phone Region"
          },
          "preset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset Id"
          },
          "preset_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset Version"
          },
          "preferred_action": {
            "type": "string",
            "title": "Preferred Action"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "can_edit",
          "version",
          "text",
          "expires_at",
          "status",
          "title",
          "categories",
          "phone_region",
          "preset_id",
          "preset_version",
          "preferred_action"
        ],
        "title": "SourceView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VersionRef": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "source_revision_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source Revision Id"
          },
          "decision_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Decision Version"
          },
          "settings_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Settings Version"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "source_revision_id",
          "decision_version",
          "settings_version"
        ],
        "title": "VersionRef"
      },
...
    }
  }
}